requirements.txt
biopython>=1.81
pandas>=2.0.1
numpy>=1.24.3
joblib>=1.2.0

